// src/frontend/components/Types.ts
export interface Track {
    id: string
    name: string
    image: string
    description: string
    audio: string
    title: string
    albumTitle: string
    artist: string
    bpm: string
    key: string
    releaseDate: string
    previewTimeStamp: string
    explicit: string
    genre: string
    duration: string
}